Fix OutputStream's async
