+++ src/auto/output_stream.rs
@@ -47,8 +47,6 @@ pub trait OutputStreamExt {
 
     fn splice<'a, P: IsA<InputStream>, Q: Into<Option<&'a Cancellable>>>(&self, source: &P, flags: OutputStreamSpliceFlags, cancellable: Q) -> Result<isize, Error>;
 
-    fn splice_async<'a, P: IsA<InputStream>, Q: Into<Option<&'a Cancellable>>, R: Fn(Result<(), Error>) + Send + Sync + 'static>(&self, source: &P, flags: OutputStreamSpliceFlags, io_priority: i32, cancellable: Q, callback: R);
-
     //#[cfg(any(feature = "v2_40", feature = "dox"))]
     //fn vprintf<'a, P: Into<Option<&'a Cancellable>>>(&self, cancellable: P, error: &mut Error, format: &str, args: /*Unknown conversion*//*Unimplemented*/Unsupported) -> Option<usize>;
 
@@ -59,11 +57,7 @@ pub trait OutputStreamExt {
     #[cfg(any(feature = "v2_44", feature = "dox"))]
     fn write_all_async<'a, P: Into<Option<&'a Cancellable>>, Q: Fn(Result<usize, Error>) + Send + Sync + 'static>(&self, buffer: &[u8], io_priority: i32, cancellable: P, callback: Q);
 
-    fn write_async<'a, P: Into<Option<&'a Cancellable>>, Q: Fn(Result<(), Error>) + Send + Sync + 'static>(&self, buffer: &[u8], io_priority: i32, cancellable: P, callback: Q);
-
     fn write_bytes<'a, P: Into<Option<&'a Cancellable>>>(&self, bytes: &glib::Bytes, cancellable: P) -> Result<isize, Error>;
-
-    fn write_bytes_async<'a, P: Into<Option<&'a Cancellable>>, Q: Fn(Result<(), Error>) + Send + Sync + 'static>(&self, bytes: &glib::Bytes, io_priority: i32, cancellable: P, callback: Q);
 }
 
 impl<O: IsA<OutputStream>> OutputStreamExt for O {
@@ -176,27 +170,6 @@ impl<O: IsA<OutputStream>> OutputStreamExt for O {
         }
     }
 
-    fn splice_async<'a, P: IsA<InputStream>, Q: Into<Option<&'a Cancellable>>, R: Fn(Result<(), Error>) + Send + Sync + 'static>(&self, source: &P, flags: OutputStreamSpliceFlags, io_priority: i32, cancellable: Q, callback: R) {
-        let cancellable = cancellable.into();
-        let cancellable = cancellable.to_glib_none();
-        let user_data: Box<Box<Fn(Result<(), Error>) + Send + Sync + 'static>> = Box::new(Box::new(callback));
-        extern "C" fn splice_async_trampoline(_source_object: *mut gobject_ffi::GObject, res: *mut ffi::GAsyncResult, user_data: glib_ffi::gpointer)
-        {
-            callback_guard!();
-            unsafe {
-                let mut error = ptr::null_mut();
-                let _ = ffi::g_output_stream_splice_finish(_source_object as *mut _, res, &mut error);
-                let result = if error.is_null() { Ok(()) } else { Err(from_glib_full(error)) };
-                let callback: &&(Fn(Result<(), Error>) + Send + Sync + 'static) = transmute(user_data);
-                callback(result);
-            }
-        }
-        let callback = splice_async_trampoline;
-        unsafe {
-            ffi::g_output_stream_splice_async(self.to_glib_none().0, source.to_glib_none().0, flags.to_glib(), io_priority, cancellable.0, Some(callback), Box::into_raw(user_data) as *mut _);
-        }
-    }
-
     //#[cfg(any(feature = "v2_40", feature = "dox"))]
     //fn vprintf<'a, P: Into<Option<&'a Cancellable>>>(&self, cancellable: P, error: &mut Error, format: &str, args: /*Unknown conversion*//*Unimplemented*/Unsupported) -> Option<usize> {
     //    unsafe { TODO: call ffi::g_output_stream_vprintf() }
@@ -249,28 +222,6 @@ impl<O: IsA<OutputStream>> OutputStreamExt for O {
         }
     }
 
-    fn write_async<'a, P: Into<Option<&'a Cancellable>>, Q: Fn(Result<(), Error>) + Send + Sync + 'static>(&self, buffer: &[u8], io_priority: i32, cancellable: P, callback: Q) {
-        let cancellable = cancellable.into();
-        let cancellable = cancellable.to_glib_none();
-        let count = buffer.len() as usize;
-        let user_data: Box<Box<Fn(Result<(), Error>) + Send + Sync + 'static>> = Box::new(Box::new(callback));
-        extern "C" fn write_async_trampoline(_source_object: *mut gobject_ffi::GObject, res: *mut ffi::GAsyncResult, user_data: glib_ffi::gpointer)
-        {
-            callback_guard!();
-            unsafe {
-                let mut error = ptr::null_mut();
-                let _ = ffi::g_output_stream_write_finish(_source_object as *mut _, res, &mut error);
-                let result = if error.is_null() { Ok(()) } else { Err(from_glib_full(error)) };
-                let callback: &&(Fn(Result<(), Error>) + Send + Sync + 'static) = transmute(user_data);
-                callback(result);
-            }
-        }
-        let callback = write_async_trampoline;
-        unsafe {
-            ffi::g_output_stream_write_async(self.to_glib_none().0, buffer.to_glib_none().0, count, io_priority, cancellable.0, Some(callback), Box::into_raw(user_data) as *mut _);
-        }
-    }
-
     fn write_bytes<'a, P: Into<Option<&'a Cancellable>>>(&self, bytes: &glib::Bytes, cancellable: P) -> Result<isize, Error> {
         let cancellable = cancellable.into();
         let cancellable = cancellable.to_glib_none();
@@ -278,27 +229,6 @@ impl<O: IsA<OutputStream>> OutputStreamExt for O {
             let mut error = ptr::null_mut();
             let ret = ffi::g_output_stream_write_bytes(self.to_glib_none().0, bytes.to_glib_none().0, cancellable.0, &mut error);
             if error.is_null() { Ok(ret) } else { Err(from_glib_full(error)) }
-        }
-    }
-
-    fn write_bytes_async<'a, P: Into<Option<&'a Cancellable>>, Q: Fn(Result<(), Error>) + Send + Sync + 'static>(&self, bytes: &glib::Bytes, io_priority: i32, cancellable: P, callback: Q) {
-        let cancellable = cancellable.into();
-        let cancellable = cancellable.to_glib_none();
-        let user_data: Box<Box<Fn(Result<(), Error>) + Send + Sync + 'static>> = Box::new(Box::new(callback));
-        extern "C" fn write_bytes_async_trampoline(_source_object: *mut gobject_ffi::GObject, res: *mut ffi::GAsyncResult, user_data: glib_ffi::gpointer)
-        {
-            callback_guard!();
-            unsafe {
-                let mut error = ptr::null_mut();
-                let _ = ffi::g_output_stream_write_bytes_finish(_source_object as *mut _, res, &mut error);
-                let result = if error.is_null() { Ok(()) } else { Err(from_glib_full(error)) };
-                let callback: &&(Fn(Result<(), Error>) + Send + Sync + 'static) = transmute(user_data);
-                callback(result);
-            }
-        }
-        let callback = write_bytes_async_trampoline;
-        unsafe {
-            ffi::g_output_stream_write_bytes_async(self.to_glib_none().0, bytes.to_glib_none().0, io_priority, cancellable.0, Some(callback), Box::into_raw(user_data) as *mut _);
         }
     }
 }
