+++ src/lib.rs
@@ -26,6 +26,7 @@ mod file;
 mod file_i_o_stream;
 mod file_input_stream;
 mod file_output_stream;
+mod output_stream;
 mod resource;
 
 pub use glib::{
@@ -55,6 +56,7 @@ pub mod prelude {
     pub use file_input_stream::FileInputStreamExt;
     pub use file_output_stream::FileOutputStreamExt;
     pub use input_stream::InputStreamExtManual;
+    pub use output_stream::OutputStreamExtManual;
 }
 
 pub use prelude::*;
